A complete run of all the stanzas on the website has been made, errors.json reflects those of Title xxxx that could not be parsed at all, At the bare minimum we should have most Title's and dates to compare to the stanza's, just need to run our config.txt through the same regexes and it should produce similar results and then we can compare the dates

+++ gen_web_stanzas.py
@@ -8,6 +8,7 @@ import os.path
 import json
 
 json_file = "complete_stanzas.json"
+error_json_file = "errors.json"
 
 
 def check_file():
@@ -23,6 +24,18 @@ def check_file():
             return data
     else:
         with open(json_file, "w") as f:
+            data = {}
+            json.dump(data, f, indent=4)
+            return data
+
+
+def check_error_file():
+    if os.path.exists(error_json_file):
+        with open(error_json_file, "r") as f:
+            data = json.load(f)
+            return data
+    else:
+        with open(error_json_file, "w") as f:
             data = {}
             json.dump(data, f, indent=4)
             return data
@@ -58,30 +71,37 @@ def gen_web_stanzas():
     web_stanzas = get_stanzas()
 
     stanza_data = check_file()
+    error_data = check_error_file()
 
-    links = get_not_updated(web_stanzas, stanza_data)[:50]
-    #print(links)
+    links = get_not_updated(web_stanzas, stanza_data)
+    # print(links)
     with Pool(10) as p:
         results = [p.apply_async(parsePage, args=(i[1],i[0])) for i in links]
         stanza_arr = [result.get() for result in results]
         for stanza in stanza_arr:
+            if len(stanza['ERROR']) > 0:
+                error_data[stanza['ERROR'][0]['link']] = stanza['ERROR']
+                with open(error_json_file, "w") as f:
+                    json.dump(error_data, f, indent=4)
+            del stanza['ERROR']
             stanza_data.update(stanza)
         with open(json_file, "w") as f:
             json.dump(stanza_data, f, indent=4)
 
-        #links = get_not_updated(all_links, stanza_data)[:5]
+        # links = get_not_updated(all_links, stanza_data)[:5]
 
 def parsePage(link, stanza_updated):
     page = requests.get(link)
     soup = BeautifulSoup(page.text, 'html.parser')
     pre_tags = soup.find_all('pre')
     
-    total_stanzas = {}
+    total_stanzas = {
+        'ERROR': []
+    }
 
     stanza_text = ""
     last_updated = ""
     title = ""
-
     for tag in pre_tags:
         lines = tag.text.split('\n')[1:]
         if "IncludeFile" in tag.text:
@@ -102,18 +122,25 @@ def parsePage(link, stanza_updated):
                         'link': link
                     }
                 else:
-                    print("Stanza without title")
+                    print("COULD NOT PARSE THE TITLE: LOGGING TO ERROR JSON")
+                    total_stanzas['ERROR'].append({
+                        'link': link, 
+                        'stanza_text': stanza_text, 
+                        'last_updated': str(last_updated),
+                        'tag_text': tag.text,
+                        'msg': "Could not parse a Title xxxxxx from this tag_text"
+                    })
                 stanza_text = ""
                 last_updated = ""
                 continue
-            if re.match(r'^Title (.+)$', line):
+            if re.match(r'^Title (.+)$', line, flags=re.I):
                 print(line)
 
                 # NOTE: Certain updates seem to be structured as Title blahblah (OCLC Include File updated xxxxxxxx)
-                search = re.search(r'^Title ((?:(?! \(updated).)*) \(updated (\d{8})\)', line)  # Wow so beautiful
+                search = re.search(r'^Title ((?:(?! \(updated).)*) \(updated (\d{8})\)', line, flags=re.I)  # Wow so beautiful
 
                 if search == None: # Make the assumption that it has no (updated) format ex: Title 123Library
-                    search = re.search(r'^Title (.+)$', line)
+                    search = re.search(r'^Title (.+)$', line, flags=re.I)
                 title = search.group(1).strip()
 
                 last_updated = ""
